Get and set default options for sbatch and sluRm internals

Most of the functions in the sluRm package use chdir and job-name options to write and submit jobs to Slurm. These options have global defaults that are set and retrieved using opts_sluRm.

opts_sluRm

Format

An object of class opts_sluRm of length 10.

Details

Current supported options are:

debug_off : function () Deactivates the debug mode.

debug_on : function () Activates the debug mode.

get_chdir : function () Gets chdir

get_cmd : function () Gets the bash command (see the Debug code section)

get_debug : function () Returns TRUE if the debug mode is on. FALSE otherwise.

get_job_name : function (check = TRUE) Returns the current value of job-name

set_chdir : function (path, recursive = TRUE) Changes chdir.

set_job_name : function (path, check = TRUE, overwrite = TRUE) Changes the job-name. When changing the name of the job the function will check whether the folder chdir/job-name is empty or not. If empty/not created it will create it, otherwise it will delete its contents (if overwrite = TRUE, else it will return with an Error).

For general set/retrieve options

get_opts : function (...)

set_opts : function (...)

Examples

# Common setup

# NOT RUN {
opts_sluRm$set_chdir("/staging/pdt/vegayon")
opts_sluRm$set_job_name("simulations-1")
opts_slurm$set_opts(partition="thomas", account="lc_pdt")
# }