Full path names for Slurm jobs

Using opts_slurmR$get_tmp_path and opts_slurmR$get_job_name creates file names with full path to the objects. This function is intended for internal use only.

snames(type, array_id = NULL, tmp_path = NULL, job_name = NULL)

Arguments

type

can be any of r, sh, out, or rds.

array_id

Integer. ID of the array to create the name.

tmp_path

Character scalar. Path to the temp directory used by the job to write files.

job_name

Character scalar. Name of the job.

Value

A character scalar. The normalized path to the corresponding file.

Details

By default, the parameters tmp_path and job_name are retrieved from the current options specified in opts_slurmR.

See also