Check for possible actions for a slurm_job wrapper

Users can choose whether to submit the job or not, to wait for it, and whether they want to collect the results right away after the job has finished. This function will help developers to figure out what set of actions need to be taken depending on the plan.

the_plan(plan)

Arguments

plan

A character scalar with either of the following values: "collect", "wait", "submit", or "none".

Value

A list with three logical scalars.

Details

This is a helper function that returns a list with three logical values: wait, collect, and submit. There are four possible cases:

plan == "collect", then all three are TRUE.

plan == "wait", then all but collect are TRUE.

plan == "submit" then only submit equals TRUE.

plan == "none" then all three are FALSE.

In general, bot wait and submit will be passed to sbatch.

When collect == TRUE, then it usually means that the function will be calling Slurm_collect right after submitting the job via sbatch.

See also

Examples

the_plan("none")
#> $collect
#> [1] FALSE
#> 
#> $wait
#> [1] FALSE
#> 
#> $submit
#> [1] FALSE
#> 
# $collect
# [1] FALSE
#
# $wait
# [1] FALSE
#
# $submit
# [1] FALSE

the_plan("wait")
#> $collect
#> [1] FALSE
#> 
#> $wait
#> [1] TRUE
#> 
#> $submit
#> [1] TRUE
#> 
# $collect
# [1] FALSE
#
# $wait
# [1] TRUE
#
# $submit
# [1] TRUE