The Slurm version of the lapply function

The Slurm version of the lapply function

Slurm_Map(f, ..., njobs = 2L, mc.cores = 1L,
  job_name = opts_sluRm$get_job_name(),
  tmp_path = opts_sluRm$get_tmp_path(), plan = "collect",
  sbatch_opt = list(ntasks = 1L, `cpus-per-task` = mc.cores),
  rscript_opt = list(vanilla = TRUE), seeds = NULL, compress = TRUE,
  export = NULL, libPaths = .libPaths(), hooks = NULL)

Slurm_lapply(X, FUN, ..., njobs = 2L, mc.cores = 1L,
  job_name = opts_sluRm$get_job_name(),
  tmp_path = opts_sluRm$get_tmp_path(), plan = "collect",
  sbatch_opt = list(ntasks = 1L, `cpus-per-task` = mc.cores),
  rscript_opt = list(vanilla = TRUE), seeds = NULL, compress = TRUE,
  export = NULL, libPaths = .libPaths(), hooks = NULL)

Slurm_sapply(X, FUN, ..., simplify = TRUE)

Arguments

njobs
Integer. Number of jobs to use in the job-array.

job_name
Character. Name of the job to be passed to Slurm.

tmp_path
Character. Path to the directory where all the data (including scripts) will be stored. Notice that this path must be accessible by all the nodes in the network (See opts_sluRm).

plan
A character scalar. (See the_plan).

sbatch_opt, rscript_opt
List. Options to be passed via flags to the bash file as #SBATCH and to Rscript respectively.

seeds
Integer vector of length njobs. Seeds to be passed to each job. When NULL (default), seeds will be picked automatically (see new_rscript).

compress
Logical scalar (default TRUE). Passed to saveRDS. Setting this value to FALSE can be useful when the user requires faster read/write of R objects on disk.

export
A named list with objects to be included in the Spawned sessions.

libPaths
A character vector. See .libPaths.

hooks
A list of functions (passed to new_slurm_job).

X, FUN, f, mc.cores, ...
Arguments passed to either parallel::mclapply or parallel::mcMap.

simplify
Logical scalar. See sapply.

Details

The function Slurm_lapply will submit njobs to the queue and distribute X according to parallel::splitIndices. For example, if X is list with 1,000 elements, and njobs = 2, then Slurm_lapply will submit 2 jobs with 500 elements of X each (2 chunks of data). The same principle applies to Slurm_sapply and Slurm_Map, this is, the data is split by chunks so all the information is sent at once when the job is submitted.

Just like sapply is to lapply, Slurm_sapply is just a wrapper of Slurm_lapply with an extra argument, simplify. When TRUE, once the job is collected, the function simplify2array is called.

References

Job Array Support https://slurm.schedmd.com/job_array.html

See also

Examples

# NOT RUN {
  # A job drawing 1e6 uniforms on 10 jobs (array)
  # The option wait=TRUE makes it return only once the job is completed.
  job1 <- Slurm_lapply(1:20, function(i) runif(1e6), njobs=10, wait = TRUE)

  # To collect
  ans <- Slurm_collect(job1)

  # As before, but this time not waiting, and now we are passing more
  # arguments to the function
  job1 <- Slurm_lapply(1:20, function(i, a) runif(1e6, a), a = -1, njobs=10,
      wait = FALSE)

  # We submit
  job1 <- sbatch(job1)

  # In order to cancel a job
  scancel(job1)

  # How to clean up
  Slurm_clean(job1)
# }