The Slurm version of the lapply function

The Slurm version of the lapply function

Slurm_lapply(X, FUN, ..., nodes = 2, mc.cores = getOption("mc.cores", 2L),
  job_name = getOption("sluRm.job_name", "sluRm"), job_path = NULL,
  submit = TRUE)

Arguments

X, FUN, mc.cores, ...
Arguments passed to parallel::mclapply.

nodes
Integer. Number of nodes to specity.

job_name
Character. Name of the job to be passed to `Slurm`.

job_path
Character. Path to the directory where all the data (including scripts) will be stored (See [options_sluRm]).

submit
Logical, when TRUE calls sbatch to submit the job to slurm.

Examples

# NOT RUN {
  Slurm_lapply(1:20, mean)
# }