R wrappers for Slurm commands

R wrappers for Slurm commands

slurm_available()

last_submitted_job()

sbatch(x, wait = FALSE, submit = TRUE, ...)

# S3 method for slurm_job
sbatch(x, wait = FALSE, submit = TRUE, ...)

squeue(x = NULL, ...)

# S3 method for default
squeue(x = NULL, ...)

# S3 method for slurm_job
squeue(x, ...)

scancel(x = NULL, ...)

# S3 method for default
scancel(x = NULL, ...)

# S3 method for slurm_job
scancel(x = NULL, ...)

sacct(x, ...)

# S3 method for default
sacct(x = NULL, brief = TRUE, parsable = TRUE, ...)

# S3 method for slurm_job
sacct(x, ...)

slurm.conf()

Arguments

x
Either an object of class slurm_job, or, in some cases, an integer as a Slurm jobid. Note that some functions allow passing no arguments.

wait
Logical scalar. When TRUE the function will pass the --wait flag to Slurm and set wait = TRUE in the system2 call.

submit
Logical, when TRUE calls sbatch to submit the job to slurm.

...
Further flags passed to the command line function.

brief, parsable
Logical. When TRUE, these are passed as flags directly to the command line function sacct.

Details

The function slurm_available checks whether Slurm is available in the system or not. It is usually called before calling any bash wrapper. If available, the function will return TRUE, otherwise FALSE.

The las_submitted_job function will return the latest slurm_job object that was submitted via sbatch in the current session. The last_job function is just an alias of the later. If no job has been submitted, then the resulting value will be NULL.

In the case of sbatch, function takes an object of class slurm_job and submits it to the queue. In debug mode the job will be submitted via sh instead.

The function slurm.conf is a wrapper of the function scontrol that returns configuration info about Slurm, in particular, the underlying command that is called is scontrol show conf. This returns a named character vector with configuration info about the cluster. The name of this function matches the name of the file that holds this information.

Examples

# Are we under a Slurm Cluster?
slurm_available()

#> [1] FALSE

# NOT RUN {
# The last_job function can be handy when `plan = "collect"` in a called,
# for example
job <- Slurm_lapply(1:1000, function(i) runif(100), njobs = 2, plan = "collect")

# Post collection analysis
state(last_job())
# }# NOT RUN {
# Submitting a simple job
job <- Slurm_EvalQ(sluRm::WhoAmI(), njobs = 4L, plan = "submit")

# Checking the status of the job (we can simply print)
job
state(job) # or use the state function
sacct(job) # or get more info with the sactt wrapper.

# Suppose one of the jobs is taking too long to complete (say #4)
# we can stop it and resubmit the job as follows:
scancel(job)

# Resubmitting only 4
sbatch(job, array = 4) # A new jobid will be assigned

# }

# NOT RUN {
# allows?
sconfig <- slurm.conf() # We first retrieve the info.
sconfig["MaxArraySize"]
# }