Source an R script as a Slurm job

This function sources R scripts using Slurm by creating a batch script file and submitting it via sbatch.

sourceSlurm(file, job_name = NULL, tmp_path = tempdir(),
  rscript_opt = list(vanilla = TRUE), plan = "wait", ...)

Arguments

file
Character. Path to the R script to source using Slurm.

job_name
Character. Name of the job to be passed to Slurm.

tmp_path
Character. Path to the directory where all the data (including scripts) will be stored. Notice that this path must be accessible by all the nodes in the network (See opts_slurmR).

rscript_opt
List. Options to be passed to Rscript.

plan
A character scalar. (See the_plan).

...
Further options passed to sbatch.

Value

Whatever sbatch returns.

Details

sourceSlurm checks for flags that may be included in the Slurm job file. If the R script starts with #!/bin/ or similar, then #SBATCH flags will be read from the R script and added to the Slurm job file.

Examples

# In this example we will be sourcing an R script that also has #SBATCH
# flags. Here are the contents
file <- system.file("example.R", package="slurmR")

cat(readLines(file), sep="\n")

#> #!/bin/sh
#> #SBATCH --account=lc_ggv
#> #SBATCH --time=01:00:00
#> #SBATCH --mem-per-cpu=4G
#> #SBATCH --job-name=Waiting
#> Sys.sleep(10)
#> message("done.")
#> 

# #!/bin/sh
# #SBATCH --account=lc_ggv
# #SBATCH --time=01:00:00
# #SBATCH --mem-per-cpu=4G
# #SBATCH --job-name=Waiting
# Sys.sleep(10)
# message("done.")

# We can directly submit this R script as a job by calling `sourceSlurm`.
# (of course you need Slurm to do this!)

# NOT RUN {
sourceSlurm(file)
# }

# The function will create a bash script that is used later to be submitted to
# the queue using `sbatch`. The resulting file looks something like this
# #!/bin/sh
# #SBATCH --job-name=Waiting
# #SBATCH --output=/home/vegayon/Documents/slurmR/Waiting.out
# #SBATCH --account=lc_ggv
# #SBATCH --time=01:00:00
# #SBATCH --mem-per-cpu=4G
# /usr/lib/R/bin/Rscript --vanilla /usr/local/lib/R/site-library/slurmR/example.R